kevins-place/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
databases[sqlite,postgresql]>=0.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
coincurve>=18.0.0
bech32>=1.2.0
pydantic>=2.5.0
httpx>=0.25.0
PyJWT>=2.8.0
bcrypt>=4.1.0
slowapi>=0.1.9
